Enhance Contractor Functionality and UI: Introduce a new debug endpoint in ContractorsController for task and project insights, including unique assignments and sample tasks. Refactor error handling in ProjectManagerController to provide more detailed error responses. Update FirebaseService to support conditional document updates and improve query filtering capabilities. Enhance UI components for task management, including collapsible task cards and improved progress reporting modals. Update CSS for better visual consistency and responsiveness across contractor dashboard elements.

+++ Web/ICCMS-Web/Models/ProjectDetailDto.cs
@@ -17,10 +17,10 @@ namespace ICCMS_Web.Models
         public string Description { get; set; } = string.Empty;
 
         [JsonPropertyName("budgetPlanned")]
-        public double BudgetPlanned { get; set; }
+        public decimal BudgetPlanned { get; set; }
 
         [JsonPropertyName("budgetActual")]
-        public double BudgetActual { get; set; }
+        public decimal BudgetActual { get; set; }
 
         [JsonPropertyName("status")]
         public string Status { get; set; } = string.Empty;
